--- app/src/main/java/org/dicekeys/app/fragments/RecipeFragment.kt
@@ -1,24 +1,0 @@
-package org.dicekeys.app.fragments
-
-import android.os.Bundle
-import android.view.View
-import dagger.hilt.android.AndroidEntryPoint
-import org.dicekeys.app.AppFragment
-import org.dicekeys.app.R
-import org.dicekeys.app.databinding.DicekeyFragmentBinding
-import org.dicekeys.app.databinding.RecipeFragmentBinding
-import org.dicekeys.app.openDialogDeleteDiceKey
-import org.dicekeys.app.encryption.BiometricsHelper
-import org.dicekeys.app.viewmodels.DiceKeyViewModel
-import javax.inject.Inject
-
-@AndroidEntryPoint
-class RecipeFragment: AppFragment<RecipeFragmentBinding>(R.layout.recipe_fragment) {
-
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
-
-    }
-}
