+++ app/src/main/java/org/dicekeys/app/fragments/dicekey/BackupFragment.kt
@@ -1,4 +1,4 @@
-package org.dicekeys.app.fragments
+package org.dicekeys.app.fragments.dicekey
 
 import android.os.Bundle
 import android.view.View
@@ -6,8 +6,6 @@ import dagger.hilt.android.AndroidEntryPoint
 import org.dicekeys.app.AppFragment
 import org.dicekeys.app.R
 import org.dicekeys.app.databinding.BackupFragmentBinding
-import org.dicekeys.app.databinding.DicekeyFragmentBinding
-import org.dicekeys.app.databinding.SeedSolokeyFragmentBinding
 
 @AndroidEntryPoint
 class BackupFragment: AppFragment<BackupFragmentBinding>(R.layout.backup_fragment) {
