+++ app/src/main/java/org/dicekeys/app/fragments/dicekey/SaveFragment.kt
@@ -1,0 +1,31 @@
+package org.dicekeys.app.fragments.dicekey
+
+import android.os.Bundle
+import android.view.View
+import dagger.hilt.android.AndroidEntryPoint
+import org.dicekeys.app.AppFragment
+import org.dicekeys.app.R
+import org.dicekeys.app.databinding.SaveFragmentBinding
+import org.dicekeys.app.encryption.BiometricsHelper
+import javax.inject.Inject
+
+@AndroidEntryPoint
+class SaveFragment: AppFragment<SaveFragmentBinding>(R.layout.save_fragment) {
+
+    @Inject
+    lateinit var biometricsHelper: BiometricsHelper
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+        binding.vm = getDiceKeyRootFragment().viewModel
+
+        binding.buttonSave.setOnClickListener {
+            biometricsHelper.encrypt(getDiceKeyRootFragment().viewModel.diceKey, this)
+        }
+
+        binding.buttonDelete.setOnClickListener {
+            getDiceKeyRootFragment().viewModel.remove()
+        }
+    }
+}
