+++ app/src/main/res/drawable/ic_secret.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="38dp"
+    android:height="19dp"
+    android:viewportWidth="38"
+    android:viewportHeight="19">
+  <path
+      android:pathData="M3,0C1.3432,0 0,1.3432 0,3V16C0,17.6569 1.3432,19 3,19H35C36.6569,19 38,17.6569 38,16V3C38,1.3432 36.6569,0 35,0H3ZM7.6905,12.6667C9.4394,12.6667 10.8571,11.2489 10.8571,9.5C10.8571,7.7511 9.4394,6.3334 7.6905,6.3334C5.9416,6.3334 4.5238,7.7511 4.5238,9.5C4.5238,11.2489 5.9416,12.6667 7.6905,12.6667ZM22.619,9.5C22.619,11.2489 21.2013,12.6667 19.4524,12.6667C17.7035,12.6667 16.2857,11.2489 16.2857,9.5C16.2857,7.7511 17.7035,6.3334 19.4524,6.3334C21.2013,6.3334 22.619,7.7511 22.619,9.5ZM31.2143,12.6667C32.9632,12.6667 34.3809,11.2489 34.3809,9.5C34.3809,7.7511 32.9632,6.3334 31.2143,6.3334C29.4654,6.3334 28.0476,7.7511 28.0476,9.5C28.0476,11.2489 29.4654,12.6667 31.2143,12.6667Z"
+      android:fillColor="#5576C5"
+      android:fillType="evenOdd"/>
+</vector>
