+++ app/src/main/res/drawable/ic_secret_with_arrow.xml
@@ -1,0 +1,13 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="38dp"
+    android:height="42dp"
+    android:viewportWidth="38"
+    android:viewportHeight="42">
+  <path
+      android:pathData="M18.0921,0L18.0921,17.0107L10.284,10.2957L9,11.3999L19,20L29,11.3999L27.716,10.2957L19.9079,17.0107L19.9079,0L18.0921,0Z"
+      android:fillColor="#5576C5"/>
+  <path
+      android:pathData="M3,23C1.3432,23 0,24.3431 0,26V39C0,40.6569 1.3432,42 3,42H35C36.6569,42 38,40.6569 38,39V26C38,24.3431 36.6569,23 35,23H3ZM7.6905,35.6667C9.4394,35.6667 10.8571,34.2489 10.8571,32.5C10.8571,30.7511 9.4394,29.3334 7.6905,29.3334C5.9416,29.3334 4.5238,30.7511 4.5238,32.5C4.5238,34.2489 5.9416,35.6667 7.6905,35.6667ZM22.619,32.5C22.619,34.2489 21.2013,35.6667 19.4524,35.6667C17.7035,35.6667 16.2857,34.2489 16.2857,32.5C16.2857,30.7511 17.7035,29.3334 19.4524,29.3334C21.2013,29.3334 22.619,30.7511 22.619,32.5ZM31.2143,35.6667C32.9632,35.6667 34.3809,34.2489 34.3809,32.5C34.3809,30.7511 32.9632,29.3334 31.2143,29.3334C29.4654,29.3334 28.0476,30.7511 28.0476,32.5C28.0476,34.2489 29.4654,35.6667 31.2143,35.6667Z"
+      android:fillColor="#5576C5"
+      android:fillType="evenOdd"/>
+</vector>
