+++ app/src/main/res/layout/main_dicekey_fragment.xml
@@ -5,6 +5,10 @@
 
     <data>
 
+        <variable
+            name="vm"
+            type="org.dicekeys.app.viewmodels.DiceKeyViewModel" />
+
     </data>
 
 
@@ -28,13 +32,13 @@
 
                 <Button
                     android:id="@+id/buttonLock"
+                    style="@style/Widget.MaterialComponents.Button.TextButton"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    style="@style/Widget.MaterialComponents.Button.TextButton"
-                    android:textSize="12sp"
-                    android:textColor="@color/white"
                     android:drawableTop="@drawable/ic_baseline_lock_24"
-                    android:text="@string/lock" />
+                    android:text="@string/lock"
+                    android:textColor="@color/white"
+                    android:textSize="12sp" />
 
                 <TextView
                     android:id="@+id/toolbarTitle"
@@ -42,43 +46,47 @@
                     android:layout_height="wrap_content"
                     android:layout_weight="1"
                     android:textAlignment="center"
+                    android:textAppearance="?attr/textAppearanceBody1"
                     android:textColor="@color/white"
-                    tools:text="DiceKey with Q4 in center"
                     android:textSize="18sp"
                     android:textStyle="bold"
-                    android:textAppearance="?attr/textAppearanceBody1" />
+                    tools:text="DiceKey with Q4 in center" />
 
                 <Button
                     android:id="@+id/buttonSave"
+                    style="@style/Widget.MaterialComponents.Button.TextButton"
                     android:layout_width="wrap_content"
                     android:layout_height="wrap_content"
-                    style="@style/Widget.MaterialComponents.Button.TextButton"
-                    android:textSize="12sp"
-                    android:textColor="@color/white"
                     android:drawableTop="@drawable/ic_baseline_lock_24"
-                    android:text="@string/save" />
+                    android:text="@{vm.isSaved() ? @string/saved : @string/save}"
+                    android:textColor="@color/white"
+                    android:textSize="12sp" />
 
             </LinearLayout>
 
         </com.google.android.material.appbar.AppBarLayout>
 
-        <com.google.android.material.tabs.TabLayout
-            android:id="@+id/tabLayout"
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/dicekey_host_fragment"
+            android:name="androidx.navigation.fragment.NavHostFragment"
+            android:layout_width="match_parent"
+            android:layout_height="0dp"
+            app:defaultNavHost="false"
+            app:layout_constraintBottom_toTopOf="@+id/bottom_navigation"
+            app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/appBarLayout2"
+            app:navGraph="@navigation/dicekey_nav_graph" />
+
+        <com.google.android.material.bottomnavigation.BottomNavigationView
+            android:id="@+id/bottom_navigation"
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
+            app:labelVisibilityMode="labeled"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:tabMode="fixed" />
-
-        <androidx.viewpager2.widget.ViewPager2
-            android:id="@+id/pager"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            app:layout_constraintBottom_toTopOf="@+id/tabLayout"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/appBarLayout2" />
+            app:menu="@menu/dicekey_bottom_nav" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
