+++ app/src/main/res/layout/solokey_fragment.xml
