--- app/src/main/res/menu/main_dicekey.xml
@@ -1,16 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<menu xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto">
-
-    <item
-        android:id="@+id/forget"
-        android:icon="@drawable/ic_baseline_lock_24"
-        android:title="@string/forget"
-        app:showAsAction="always"/>
-
-    <item
-        android:id="@+id/delete"
-        android:title="@string/delete"
-        app:showAsAction="never"/>
-
-</menu>
